+++ mlens/parallel/tests/test_a_single_run.py
@@ -12,16 +12,14 @@ from mlens.externals.sklearn.base import clone
 X, y = Data('stack', False, False).get_data((6, 2), 2)
 
 lc = LayerContainer()
-lc.add(ECM, 'full')
+lc.add(ECM, 'full', dtype=np.float64)
 
 def get_gt():
     """Build ground truth."""
 
     F = np.empty((X.shape[0], len(ECM)))
 
-
     for i, (_, est) in enumerate(ECM):
-
         e = clone(est)
 
         assert e is not est
